Fix(auth): Mudança na hierarquia de pastas e inclusão de APIRouter
