+++ app/models/models_auth.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+
+class UserModel(BaseModel):
+    email: str
+    password: str
+
+
